First working port to windows
using external visual studio project files
Not merged with latest changes to matgui

+++ include/matgui/matgui-common.h
@@ -26,8 +26,8 @@ namespace MatGui {
 
 constexpr double pi = 3.1415926535897932384626433832795028841971693;
 constexpr double pi2 = pi * 2;
-constexpr float pif = pi;
-constexpr float pi2f = pi2;
+constexpr float pif = (float)pi;
+constexpr float pi2f = (float)pi2;
 
 
 inline double roundDown(double value, double factor = 1.){
@@ -40,3 +40,4 @@ inline double roundMiddle(double value, double factor = 1.){
 
 } //namespace MatGui
 
+#include <ciso646>
